[IMP] base_ir_actions_report_requence: rename and make more generic
