--- base_ir_actions_report_sequence/i18n/base_ir_actions_report_sequence.pot
@@ -1,29 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* base_ir_actions_report_sequence
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 13.0\n"
-"Report-Msgid-Bugs-To: \n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: base_ir_actions_report_sequence
-#: model:ir.model,name:base_ir_actions_report_sequence.model_ir_actions_actions
-msgid "Actions"
-msgstr ""
-
-#. module: base_ir_actions_report_sequence
-#: model:ir.model,name:base_ir_actions_report_sequence.model_ir_actions_report
-msgid "Report Action"
-msgstr ""
-
-#. module: base_ir_actions_report_sequence
-#: model:ir.model.fields,field_description:base_ir_actions_report_sequence.field_ir_actions_report__priority
-msgid "Sequence"
-msgstr ""
