--- base_ir_actions_report_sequence/views/ir_actions_report_views.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!-- Copyright 2023 ForgeFlow S.L.
-     License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl). -->
-<odoo>
-    <record id="act_report_xml_view_priority" model="ir.ui.view">
-        <field name="name">ir.actions.report - base_ir_actions_report_sequence</field>
-        <field name="model">ir.actions.report</field>
-        <field name="inherit_id" ref="base.act_report_xml_view" />
-        <field name="arch" type="xml">
-            <field name="report_type" position="after">
-                <field name="priority" />
-            </field>
-        </field>
-    </record>
-    <record id="act_report_xml_view_tree_priority" model="ir.ui.view">
-        <field
-            name="name"
-        >ir.actions.report.tree - base_ir_actions_report_sequence</field>
-        <field name="model">ir.actions.report</field>
-        <field name="inherit_id" ref="base.act_report_xml_view_tree" />
-        <field name="arch" type="xml">
-            <field name="name" position="before">
-                <field name="priority" optional="show" widget="handle" />
-            </field>
-        </field>
-    </record>
-</odoo>
