+++ base_ir_actions_sequence/README.rst
@@ -1,6 +1,6 @@
-===========================
-Base Action Report Sequence
-===========================
+====================
+Base Action Sequence
+====================
 
 .. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
    !! This file is generated by oca-gen-addon-readme !!
@@ -14,10 +14,10 @@ Base Action Report Sequence
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fserver--ux-lightgray.png?logo=github
-    :target: https://github.com/OCA/server-ux/tree/13.0/base_ir_actions_report_sequence
+    :target: https://github.com/OCA/server-ux/tree/13.0/base_ir_actions_actions_sequence
     :alt: OCA/server-ux
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/server-ux-13-0/server-ux-13-0-base_ir_actions_report_sequence
+    :target: https://translation.odoo-community.org/projects/server-ux-13-0/server-ux-13-0-base_ir_actions_actions_sequence
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
     :target: https://runbot.odoo-community.org/runbot/250/13.0
@@ -36,7 +36,10 @@ they are shown.
 Usage
 =====
 
-To use this module, you need to:
+This module introduces the base to add the ability to edit the order in which
+records of models that inherit from ir.actions.actions.
+
+To use this module, you need to (example for Reports):
 
 #. Activate the developer mode
 #. Go to *Settings > Technical > Reports*.
@@ -88,6 +91,6 @@ Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
 
 |maintainer-JaumeBforgeFlow| 
 
-This module is part of the `OCA/server-ux <https://github.com/OCA/server-ux/tree/13.0/base_ir_actions_report_sequence>`_ project on GitHub.
+This module is part of the `OCA/server-ux <https://github.com/OCA/server-ux/tree/13.0/base_ir_actions_actions_sequence>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
