+++ base_ir_actions_sequence/__manifest__.py
@@ -1,13 +1,13 @@
 # Copyright 2023 ForgeFlow, S.L.
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 {
-    "name": "Base Actions Report Sequence",
+    "name": "Base Actions Sequence",
     "version": "13.0.1.0.0",
     "category": "Extra Tools",
     "author": "ForgeFlow, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/server-ux",
     "license": "LGPL-3",
     "depends": ["base"],
-    "data": ["views/ir_actions_report_views.xml"],
+    "data": ["views/actions_views.xml"],
     "installable": True,
 }
