+++ base_ir_actions_sequence/static/description/index.html
