--- setup/base_ir_actions_report_sequence/odoo/addons/base_ir_actions_report_sequence
@@ -1,1 +1,0 @@
-../../../../base_ir_actions_report_sequence
