+++ setup/base_ir_actions_sequence/odoo/addons/base_ir_actions_sequence
@@ -1,0 +1,1 @@
+../../../../base_ir_actions_sequence
